requirement.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==2.2.5
django-ckeditor==5.9.0
django-haystack==2.8.1
django-js-asset==1.2.2
django-mirage-field==1.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
djangorestframework==3.11.0
idna==2.9
jieba==0.42.1
jmespath==0.9.5
mysqlclient==1.4.6
Pillow==7.0.0
pycparser==2.20
python3-memcached==1.51
pytz==2019.3
qiniu==7.2.8
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tqdm==4.46.0
urllib3==1.25.9
Whoosh==2.7.4
